Slowly adding blending to the model.
Added code to add lc1 vectors with random elments to seed the model.
They should not bias the model in any way - the point is to allow the model to explore the space.

Using action categories "green", "orange", "around"
(should consider, "other"?)

Need to classifiy the networks action as one of those.
(that's fairly simple, because I know it almost always clicks orange)
(it does not know "around") (at least the trained network I'm using doesn't)